docs(MaskedInput): add docs for MaskedInput
