+++ docs/pages/components/input/en-US/index.md
@@ -6,6 +6,7 @@ Instead of HTML native controls, input, textarea.
 - `<InputGroup>` The input box combination component.
 - `<InputGroup.Button>` Combined with the button.
 - `<InputGroup.Addon>` Custom additional elements.
+- `<MaskedInput>` The input box component with mask.
 
 ## Import
 
@@ -49,6 +50,12 @@ Instead of HTML native controls, input, textarea.
 
 <!--{include:`tooltip.md`}-->
 
+### Masked Input
+
+MaskedInput is an input mask component. It can create input masks for phone numbers, dates, currencies, zip codes, percentages, emails, and almost anything.
+
+<!--{include:`masked-input.md`}-->
+
 ## Props
 
 ### `<Input>`
@@ -70,3 +77,15 @@ Instead of HTML native controls, input, textarea.
 | classPrefix | string `('input-group')`              | The prefix of the component CSS class   |
 | inside      | boolean                               | Sets the composition content internally |
 | size        | enum: 'lg', 'md', 'sm', 'xs' `('md')` | An input group can have different sizes |
+
+### `<MaskedInput>`
+
+`MaskedInput` extends all props of `Input`. But does not have the `type` prop.
+
+| Property          | Type `(Default)`  | Description                                                                                             |
+| ----------------- | ----------------- | ------------------------------------------------------------------------------------------------------- |
+| mask (\*)         | array, function   | Used to define how to block user input.                                                                 |
+| guide             | boolean           | In guide mode or no guide mode                                                                          |
+| placeholderChar   | string `('_')`    | The placeholder character represents the fillable spot in the mask                                      |
+| keepCharPositions | boolean `(false)` | When `true`, adding or deleting characters will not affect the position of existing characters.         |
+| showMask          | boolean           | When the input value is empty, the mask is displayed as a placeholder instead of a regular placeholder. |
