+++ docs/pages/components/input/fragments/disabled.md
@@ -13,7 +13,7 @@ const instance = (
     <InputGroup disabled style={styles}>
       <Input value="A piece of text for demonstration." />
       <InputGroup.Addon>
-        <Search />
+        <SearchIcon />
       </InputGroup.Addon>
     </InputGroup>
     <hr />
