+++ docs/pages/components/input/fragments/import.md
@@ -1,7 +1,8 @@
 ```js
-import { Input, InputGroup } from 'rsuite';
+import { Input, InputGroup, MaskedInput } from 'rsuite';
 
 // or
 import Input from 'rsuite/Input';
 import InputGroup from 'rsuite/InputGroup';
+import MaskedInput from 'rsuite/MaskedInput';
 ```
